Add defense to edit and delete post, only poster and edit and delete.

+++ game_dates/views.py
@@ -231,9 +231,13 @@ def bookingUpdateSubmit(request, booking_id):
 
 def bookingDelete(request, booking_id):
     booking = Appointment.objects.get(pk=booking_id)
-    booking.delete()
-    messages.success(request, "Event deleted!")
-    return redirect('index')
+    if booking.user == request.user:
+        booking.delete()
+        messages.success(request, "Event deleted!")
+        return redirect('index')
+    else:
+        messages.error(request, "You do not have permission to delete this Event")
+        return redirect('index')
 
 
 def checkEditTime(times, day, booking_id):
